Adding auditing system to production (#241)
* Revert "Feat: Make synth data generation more customiseable (#67)"

This reverts commit c4a1369638cec2781c9c55668d6c5b9f4e5dc342.

* added stream_metric

* added stream_metric - pt2

* Update score_results.py

* tweaked stream_time

* debugs

* corrected stream_time_init

* added checking_data column

* entry node minimal logs

* Revert "entry node minimal logs"

This reverts commit 387c0636793a8f0395e01f59c1a97af543d69399.

* pick changes from hybrid node for auditing

* adding auditing router

* lil typo

* add aiohttp

* debug

* add IS_AUDITOR to switch oor not

* debug

* debug

* debug

* debug

* debug

* debug

* update version

* debug

* debug

* last tweaks

* update docs

* add render changes

* last stuff

* add .sql for latest_scores_url table

* remove render mention

* remove server side of auditing

* cleanup

* debug file download

* add IS_AUDITOR=1 in auditor config

* new instance debug

* add doc auditing

* default to auditing

* delete duplicate file

* remove timezone

* Auditing refactors (#250)

* Feature/prod auditing (#249)

* pick changes from hybrid node for auditing

* adding auditing router

* lil typo

* add aiohttp

* debug

* add IS_AUDITOR to switch oor not

* debug

* debug

* debug

* debug

* debug

* debug

* update version

* debug

* debug

* last tweaks

* update docs

* add .sql for latest_scores_url table

* remove render mention

* remove server side of auditing

* cleanup

* debug file download

* add IS_AUDITOR=1 in auditor config

* new instance debug

* auditing refactors

* revert dev change

* fixed merge conflict

* handle keyerror

---------


* remove duplicate .sql file

* debug

* Feature/prod auditing add version (#254)

* add versions

* add versions

* cleanup

* remove mentions of checking_data

* remove mentions of checking_data

* add config[IS_AUDITOR] = '1' to auditor create config

* handled WITH_AUTOUPDATE=0 in auditor bootstrap

* cleanup docs

* corner case handling

* debug calculations.py

* upgrade bittensor in bootstrap

* fiber upgrade

* Update audit.py

---------

+++ utils/autoupdate_validator_steps.sh
@@ -2,9 +2,34 @@
 
 # THIS FILE CONTAINS THE STEPS NEEDED TO AUTOMATICALLY UPDATE THE REPO ON A TAG CHANGE
 # THIS FILE ITSELF MAY CHANGE FROM UPDATE TO UPDATE, SO WE CAN DYNAMICALLY FIX ANY ISSUES
+source $HOME/.venv/bin/activate
 
-docker image prune -a -f --filter "until=168h"
+if [ -f .vali.env ]; then
+    export $(cat .vali.env | grep -v '^#' | xargs)
+else
+    echo "Error: .vali.env file not found"
+fi
 
-docker compose --env-file .vali.env -f docker-compose.yml run -e LOCALHOST=false --entrypoint "python src/migration.py" control_node
-./utils/launch_validator.sh
+
+# docker compose --env-file .vali.env -f docker-compose.yml run -e LOCALHOST=false --entrypoint "python src/migration.py" control_node
+# ./utils/launch_validator.sh
+# task auditor-autoupdates
+
+sh -c '[ -f .auditor.env ] || cp .vali.env .auditor.env'
+
+if [ -z "${IS_AUDITOR}" ] || [ "${IS_AUDITOR}" = "1" ]; then
+    docker compose --env-file .vali.env --profile entry_node_profile down --remove-orphans
+    docker image prune -a -f --filter "until=168h"
+    
+    pip install -e .
+    pip install -r validator/control_node/requirements.txt
+
+    source $HOME/.venv/bin/activate
+
+    pm2 delete auditor || true
+    pm2 start auditor-ecosystem.config.js
+else
+    ./utils/launch_validator.sh
+fi
+
 echo "Autoupdate steps complete :-)"
